+++ mex/editor/client/src/app/shared/backend-proxy.service.ts
@@ -1,9 +1,9 @@
 import { HttpClient, httpResource } from "@angular/common/http";
 import { inject, Injectable, type Signal } from "@angular/core";
+import type { Observable } from "rxjs";
 
 import type { PreviewOrganizationalUnit } from "../shared/models/organizational-unit";
 import type { PaginatedItemsContainer } from "./models/paginated-items-container";
-import type { Observable } from "rxjs";
 
 @Injectable({ providedIn: "root" })
 /**